Pipeline Upgrades!
A lot of small changes have been made to function headers, test classes,
etc. However, the most prominent changes have all been local to the
Screen class. Namely, the 2 biggest contenders are the addition of fully
optimized 32-bit multi-translucent color composition and a very early
implementation of arbitrary sprite scaling (a feature which I didn't
initially plan to support at all)!

Translucent sprites can now be blended to a translucent background; the
blending operation is also even faster now, substituting the division by
255 when normalizing components in favor of an 8-bit right shift. The
precision discrepancy between the 2 operations is theoretically
negligible and, in practice, totally non-existent!

Sprites can now be scaled uniformly about each axis (perhaps arbitrary
aspect ratios will come in the future for both scaling and sprites in
general!)! The algorithm still needs a little optimizing; in some
scenarios little improvement can be made. There's always the possibility
it happens such that users will simply need to take care when pushing
many scaled sprites through the pipeline, but it's definitely worth
looking over a few times.

More optimizations are planned for the pixel pipeline in general, and
while some will be implemented soon (such as condition-less depth
checking), most of them are non-essential and inconvenient to work
around. As a result, many optimizations won't be implemented until the
entire rendering pipeline is finalized or when and if I decide to
support an official release of this project- whichever comes first. As
of right now, these include:

-inlining certain methods, such as compose(), at various hot call sites
-eliminating the switch statement in the outer loop of the drawing
methods in favor of variable loop directions and/ or a small lookup
table containing the possible transformations

Last but certainly worth mentioning, clear(int) has been replaced by
refresh(long) which, dependent on bit 32 in the argument passed, either
returns the "read-from" BufferedImage containing the previous frame
which has been blended with the contents of the current "written-to"
frame and clears (or blends) the contents of the read-from frame with
the 4 least significant bytes of the argument passed, OR if the user so
chooses to manage their own clearing (blitting a background sprite for
example), a BufferedImage encapsulating the contents of the written-to
pixels.